Move slow real-compile e2e tests under tests/e2e/slow/
