+++ tests/e2e/slow/__init__.py
@@ -1,0 +1,1 @@
+"""Slow e2e tests (real toolchain compiles); excluded from default e2e runs."""
