+++ tests/e2e/slow/esp32/__init__.py
@@ -1,0 +1,1 @@
+"""Slow ESP32 e2e tests (real toolchain compiles)."""
