+++ tests/e2e/slow/libretiny/__init__.py
@@ -1,0 +1,1 @@
+"""Slow LibreTiny e2e tests (real toolchain compiles)."""
